+++ packages/nuxt/test/fixtures/nuxt/app/pages/other.vue
@@ -1,0 +1,4 @@
+<template>
+  <label>Typing sample <input id="typing-sample" /></label>
+  <NuxtLink to="/">First page</NuxtLink>
+</template>